requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
httpx>=0.24.0
pydantic>=1.10.7,<2.0.0
supabase>=1.0.3
twilio>=8.1.0
google-cloud-aiplatform>=1.25.0
python-multipart>=0.0.6
ShopifyAPI>=12.2.0
pytest>=7.3.1
python-jose>=3.3.0
